+++ example/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:media_projection_screenshot/screenshot.dart';
+import 'package:media_projection_screenshot/media_projection_screenshot.dart';
 
 void main() {
   runApp(const MyApp());
@@ -14,7 +14,7 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  final _screenshotPlugin = Screenshot();
+  final _screenshotPlugin = MediaProjectionScreenshot();
 
   Uint8List? bytes;
 
@@ -33,22 +33,27 @@ class _MyAppState extends State<MyApp> {
         body: Center(
           child: Column(
             children: [
-              bytes != null
-                  ? Image.memory(
-                      bytes!,
-                      width: double.infinity,
-                      fit: BoxFit.contain,
-                    )
-                  : Container(),
               MaterialButton(
                 child: const Text('Take Capture'),
                 onPressed: () async {
-                  Uint8List? result = await _screenshotPlugin.takeCapture(x: 0, y: 0, width: 1080, height: 100);
+                  Uint8List? result = await _screenshotPlugin.takeCapture();
                   setState(() {
                     bytes = result;
                   });
                 },
               ),
+              bytes != null
+                  ? Container(
+                      padding: const EdgeInsets.all(6),
+                      color: Colors.blueAccent,
+                      height: 600,
+                      child: Image.memory(
+                        bytes!,
+                        width: double.infinity,
+                        fit: BoxFit.contain,
+                      ),
+                    )
+                  : Container(),
             ],
           ),
         ),